Separate integration file into integration_testing modules
