+++ actionpack/lib/action_dispatch/http/response.rb
@@ -19,8 +19,8 @@ module ActionDispatch # :nodoc:
   # Nevertheless, integration tests may want to inspect controller responses in
   # more detail, and that's when \Response can be useful for application
   # developers. Integration test methods such as
-  # ActionDispatch::Integration::Session#get and
-  # ActionDispatch::Integration::Session#post return objects of type
+  # ActionDispatch::IntegrationTesting::Session#get and
+  # ActionDispatch::IntegrationTesting::Session#post return objects of type
   # TestResponse (which are of course also of type \Response).
   #
   # For example, the following demo integration test prints the body of the
