+++ actionpack/lib/action_dispatch/testing/test_response.rb
@@ -1,8 +1,8 @@
 require "action_dispatch/testing/request_encoder"
 
 module ActionDispatch
-  # Integration test methods such as ActionDispatch::Integration::Session#get
-  # and ActionDispatch::Integration::Session#post return objects of class
+  # Integration test methods such as ActionDispatch::IntegrationTesting::Session#get
+  # and ActionDispatch::IntegrationTesting::Session#post return objects of class
   # TestResponse, which represent the HTTP response results of the requested
   # controller actions.
   #
